project/requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
pillow==10.4.0
